Show current speed

+++ app/src/main/java/com/egoriku/grodnoroads/domain/model/UserPosition.kt
@@ -5,13 +5,13 @@ import com.google.android.gms.maps.model.LatLng
 data class UserPosition(
     val latLng: LatLng,
     val bearing: Float,
-    val speed: Double
+    val speed: Int
 ) {
     companion object {
         val None = UserPosition(
             latLng = LatLng(0.0, 0.0),
             bearing = 0f,
-            speed = 0.0
+            speed = 0
         )
     }
 }
